requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning & NLP
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
joblib>=1.3.0
pyarrow>=15.0.2

# OpenAI API
openai>=1.0.0

# Data validation
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Excel file handling
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Optional: For better performance
accelerate>=0.20.0